move ad-hoc variables to end of legacy variables
This avoids shifting the dataset variables hardcoded by position (integers,
tags, booleans, media items), which obviously confuses all the filtering and
dimensional charting code.

Once the management of variables is refactored this will not be relevant
anymore, but until then we need to reconduce the dataset used to the shape
expected by the legacy code.