add support for custom Pattrn-core variable names
This commit introduces the ability to map dataset-specific variable
names to the core field names that are expected for the Pattrn core
code.

For example, some datasets stored in Google Sheets for demo platforms
created in 2015 used field names such as geo_latitude, geo_longitude
instead of those defined in the Pattrn code (latitude, longitude for
these two core variables).

In order to be able to keep using these datasets without changing the
datasets themselves, we add here the ability to describe core Pattrn
variables in a new schema version (2) of the recently introduced
JSON metadata document for Pattrn datasets.