re-add event counter
Needs fixing: the legacy code was only suitable for a single counter but
this would now get executed N times when N layer groups are defined.

Moreover, even with a single layer group the legacy code will not work
because of clashes with other code introduced in the meanwhile.

But we're putting this back in for the moment, to be fixed within the
2.0.0 beta cycle.